refactor: move cli path utils package
